# springdoc-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/project/{projectId}/documents/{projectDocumentId}/upload": {
      "post": {
        "tags": [
          "Project Controller"
        ],
        "summary": "Upload a new version of a project document",
        "description": "Uploads a new version of an existing project document, optionally including a change note for version tracking.",
        "operationId": "uploadNewVersion",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "projectDocumentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "customerId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "changeNote",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperString"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/create": {
      "post": {
        "tags": [
          "Project Controller"
        ],
        "summary": "Create a new project",
        "description": "Creates a new project using the provided details such as name, description, and customer information.",
        "operationId": "createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperProjectResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/assign": {
      "post": {
        "tags": [
          "Project Controller"
        ],
        "summary": "Assign a user to a project with a role",
        "description": "Assigns a specific user to a given project and associates them with a defined role (e.g., manager, developer).",
        "operationId": "assignUserToProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignUserToProjectRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperString"
                }
              }
            }
          }
        }
      }
    },
    "/api/permission/create": {
      "post": {
        "tags": [
          "Permissions"
        ],
        "summary": "Create a new permission",
        "description": "Creates a new permission entry using the provided request data.",
        "operationId": "createPermission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperPermissionResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/customer/create": {
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Create a new customer",
        "description": "Creates a new customer using the provided details.",
        "operationId": "createCustomer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperCustomerResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/update": {
      "patch": {
        "tags": [
          "User Controller"
        ],
        "summary": "Update an existing user",
        "description": "Applies partial updates to an existing user based on the provided data.",
        "operationId": "applyGeneralUpdates",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperUserResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/customer/update": {
      "patch": {
        "tags": [
          "Customers"
        ],
        "summary": "Update an existing customer",
        "description": "Updates one or more fields of an existing customer record.",
        "operationId": "applyGeneralUpdates_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperCustomerResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/customer/all": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Get all customers with pagination",
        "description": "Retrieves a paginated list of customers. Default page=0, size=20.",
        "operationId": "getAllCustomers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperListCustomerResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/project/delete/{projectId}": {
      "delete": {
        "tags": [
          "Project Controller"
        ],
        "summary": "Delete a project",
        "description": "Deletes a project permanently based on the given project ID.",
        "operationId": "deleteProject",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperString"
                }
              }
            }
          }
        }
      }
    },
    "/api/permission/delete/{id}": {
      "delete": {
        "tags": [
          "Permissions"
        ],
        "summary": "Delete a permission",
        "description": "Deletes a specific permission identified by its ID. Set forceDelete=true if needed.",
        "operationId": "deletePermission",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperPermissionResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/customer/delete/{customerId}": {
      "delete": {
        "tags": [
          "Customers"
        ],
        "summary": "Delete a customer",
        "description": "Deletes a customer by ID. Optionally, set forceDelete=true to remove related records.",
        "operationId": "deleteCustomer",
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "forceDelete",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperString"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiResponseWrapperUserResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/UserResponseDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserResponseDTO": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseWrapperString": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateProjectRequestDTO": {
        "required": [
          "customerId",
          "standardId"
        ],
        "type": "object",
        "properties": {
          "customerId": {
            "type": "integer",
            "format": "int32"
          },
          "standardId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponseWrapperProjectResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ProjectResponseDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProjectDocumentDTO": {
        "type": "object",
        "properties": {
          "projectDocumentId": {
            "type": "integer",
            "format": "int32"
          },
          "templateId": {
            "type": "integer",
            "format": "int32"
          },
          "fileId": {
            "type": "integer",
            "format": "int32"
          },
          "statusId": {
            "type": "integer",
            "format": "int32"
          },
          "versionNumber": {
            "type": "integer",
            "format": "int32"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProjectResponseDTO": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "integer",
            "format": "int32"
          },
          "customerId": {
            "type": "integer",
            "format": "int32"
          },
          "standardId": {
            "type": "integer",
            "format": "int32"
          },
          "statusId": {
            "type": "integer",
            "format": "int32"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "completionDate": {
            "type": "string",
            "format": "date-time"
          },
          "projectDocuments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectDocumentDTO"
            }
          }
        }
      },
      "AssignUserToProjectRequestDTO": {
        "required": [
          "projectId",
          "userId"
        ],
        "type": "object",
        "properties": {
          "projectId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "roleId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CreatePermissionRequestDTO": {
        "required": [
          "permissionName"
        ],
        "type": "object",
        "properties": {
          "permissionName": {
            "maxLength": 50,
            "minLength": 0,
            "pattern": "^[A-Z_]+$",
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ApiResponseWrapperPermissionResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PermissionResponseDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PermissionResponseDTO": {
        "type": "object",
        "properties": {
          "permissionId": {
            "type": "integer",
            "format": "int32"
          },
          "permissionName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateCustomerRequestDTO": {
        "required": [
          "email",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "email": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ApiResponseWrapperCustomerResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/CustomerResponseDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CustomerResponseDTO": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PatchFieldBoolean": {
        "type": "object",
        "properties": {
          "provided": {
            "type": "boolean"
          }
        }
      },
      "PatchFieldInteger": {
        "type": "object",
        "properties": {
          "provided": {
            "type": "boolean"
          }
        }
      },
      "PatchFieldString": {
        "type": "object",
        "properties": {
          "provided": {
            "type": "boolean"
          }
        }
      },
      "UpdateUserRequestDTO": {
        "type": "object",
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/PatchFieldInteger"
          },
          "firstName": {
            "$ref": "#/components/schemas/PatchFieldString"
          },
          "lastName": {
            "$ref": "#/components/schemas/PatchFieldString"
          },
          "isActive": {
            "$ref": "#/components/schemas/PatchFieldBoolean"
          }
        }
      },
...
      "UpdateCustomerRequestDTO": {
        "type": "object",
        "properties": {
          "customerId": {
            "$ref": "#/components/schemas/PatchFieldInteger"
          },
          "name": {
            "$ref": "#/components/schemas/PatchFieldString"
          },
          "email": {
            "$ref": "#/components/schemas/PatchFieldString"
          }
        }
      },
...
      "ApiResponseWrapperListCustomerResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerResponseDTO"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}